+++ crates/solarxy-core/src/aabb.rs
@@ -1,6 +1,6 @@
 use cgmath::{InnerSpace, Point3, Vector3};
 
-#[derive(Clone, Copy)]
+#[derive(Debug, Clone, Copy)]
 pub struct AABB {
     pub min: Point3<f32>,
     pub max: Point3<f32>,